Changes constant type to better match needs

+++ piwigo/Supporting Files/Macros/ImagesCollection.m
@@ -9,19 +9,19 @@
 #import "ImagesCollection.h"
 #import "Model.h"
 
-NSInteger const kAlbumCellSpacing = 8;              // Spacing between albums (horizontally and vertically)
-NSInteger const kAlbumMarginsSpacing = 4;           // Left and right margins for albums
+CGFloat const kAlbumCellSpacing = 8;              // Spacing between albums (horizontally and vertically)
+CGFloat const kAlbumMarginsSpacing = 4;           // Left and right margins for albums
 
-NSInteger const kImageCellSpacing4iPhone = 1;       // Spacing between images (horizontally and vertically)
-NSInteger const kImageCellHorSpacing4iPad = 8;
-NSInteger const kImageCellHorSpacing4iPadPopup = 1;
-NSInteger const kImageCellVertSpacing4iPad = 8;
-NSInteger const kImageCellVertSpacing4iPadPopup = 1;
-NSInteger const kImageMarginsSpacing = 4;           // Left and right margins for images
-NSInteger const kThumbnailFileSize = 144;           // Default Piwigo thumbnail file size
+CGFloat const kImageCellSpacing4iPhone = 1;       // Spacing between images (horizontally and vertically)
+CGFloat const kImageCellHorSpacing4iPad = 8;
+CGFloat const kImageCellHorSpacing4iPadPopup = 1;
+CGFloat const kImageCellVertSpacing4iPad = 8;
+CGFloat const kImageCellVertSpacing4iPadPopup = 1;
+CGFloat const kImageMarginsSpacing = 4;           // Left and right margins for images
+CGFloat const kThumbnailFileSize = 144;           // Default Piwigo thumbnail file size
 
-NSInteger const kImageDetailsCellSpacing = 8;       // Spacing between image details cells
-NSInteger const kImageDetailsMarginsSpacing = 16;   // Left and right margins for image details cells
+CGFloat const kImageDetailsCellSpacing = 8;       // Spacing between image details cells
+CGFloat const kImageDetailsMarginsSpacing = 16;   // Left and right margins for image details cells
 
 @implementation ImagesCollection
